OKEGui: if noNuma is set, don't set numa node or thread affinity
When "skip numa detection" is enabled, we shouldn't start all jobs
on node 0.

while we're at it, also set default thread affinity to 64 processors
on a node (instead of only 28). Otherwise if you have 64 processors
per numa node and you start two jobs on each node, the two jobs will
compete for 28 of the 64 processors.

Signed-off-by: akarin <i@akarin.info>